+++ package.json
@@ -17,7 +17,11 @@
   },
   "devDependencies": {
     "@testing-library/react": "^9.1.3",
-    "eslint": "^6.5.1"
+    "@types/react": "^16.9.15",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-router-dom": "^5.1.3",
+    "eslint": "^6.5.1",
+    "typescript": "^3.7.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -27,7 +31,7 @@
   },
   "jest": {
     "collectCoverageFrom": [
-      "src/**/**/*.jsx",
+      "src/**/**/*.{jsx,tsx}",
       "!src/index.js",
       "!src/firebase.prod.js",
       "!src/hooks/*.js",
